更改rating类型为Double

+++ RedShirt/Employee.swift
@@ -12,7 +12,7 @@ import RealmSwift
 class EmployeeModel:Object{
     dynamic var name:String = ""
     dynamic var brief:String = ""
-    dynamic var rating:Int = 0
+    dynamic var rating:Double = 0.0
     dynamic var portrait:NSData = NSData()
     
 }
